+++ promviz/src/promviz/SRTMDEM.java
@@ -28,7 +28,7 @@ public class SRTMDEM extends DEMFile {
 		return new LittleEndianDataInputStream(new BufferedInputStream(new FileInputStream(path)));
 	}
 	
-	public double getNextSample(Object reader) throws IOException {
+	public float getNextSample(Object reader) throws IOException {
 		int elev = ((LittleEndianDataInputStream)reader).readShort();
 		if (elev == -32768) {
 			elev = 0; // cgiar has voids filled so nodata is actually ocean
